move validation helper methods to separate module
